Restored postcontroller.py because posts_controller.py was contrived and had unnecessary code. The user will not have the ability to delete or change a post once they add it.
In post.py, I removed the updated_at column because of the previous reasons.
I added time_limit because the posts will have a limit on the amount of
time they will be in the system. The default time is 60 (I guess minutes right
now but this can change).

I appreciate how Robert changed the layout of the data base. However,
I don't think rewriting code that already exists is an effective
team player thing to do, especially  when the code is harder to read
and has functionality that is not relevant to the project.